refactor(meta_rebind_value): move to meta_rebind
